backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# OpenAI compatible client
openai==1.57.4
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# File handling
aiofiles==24.1.0
Pillow==11.0.0

# For PDF export
reportlab==4.2.5
markdown==3.7
